# apiDoc.yaml
openapi: 3.0.1
info:
  title: Personnal Budget 2
  version: 1.5.0
  description: >-
    Version 2 of the personnal budget app created for codecademy back-end
    lessons
paths:
  /enveloppes:
    summary: Get all budget enveloppes
    description: The defult path to view all budget enveloppes
    get:
      summary: Get all the enveloppes
      description: Use this to view all the budgets
      operationId: get_all_enveloppes
      responses:
        '200':
          description: Retrieve all the enveloppes
          content:
            application/json:
              examples:
                Success:
                  value:
                    enveloppes:
                      - id: 1
                        name: Main
                        initialAmount: 1000
                        parent: null
                        operationAmount: 300
                      - id: 2
                        name: Hobbies
                        initialAmount: 300
                        parent: 1
                        operationAnmount: 42
        '500':
          description: Internal server Error
      tags:
        - View all
    post:
      summary: Create new enveloppe
      description: Create new enveloppe
      operationId: post_new_enveloppe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/enveloppe'
      responses:
        '201':
          description: successfully created enveloppe
          content:
            application/json:
              examples:
                Success:
                  value:
                    enveloppes:
                      id: 1
                      name: name
                      initialAmount: 1000
                      parent: 0
                      operatienAmount: 0
        '404':
          description: Enveloppe not found
        '500':
          description: Internal server Error
      tags:
        - One enveloppe operation
...
components:
  schemas:
    enveloppe:
      type: object
      properties:
        name:
          type: string
        initialAmount:
          type: integer
        parent: 
          type: integer
...